feat: refactored settings page 👨‍💻
