+++ src/components/Settings/components/Leftside/SettingsItem.vue
@@ -1,3 +1,7 @@
+<!--
+  Copyright D3 Ledger, Inc. All Rights Reserved.
+  SPDX-License-Identifier: Apache-2.0
+-->
 <template>
   <el-row class="settings_item">
     <div class="settings_item-header">
