+++ src/components/Settings/components/Leftside/TimeZoneItem.vue
@@ -1,3 +1,7 @@
+<!--
+  Copyright D3 Ledger, Inc. All Rights Reserved.
+  SPDX-License-Identifier: Apache-2.0
+-->
 <template>
   <SettingsItem
     :v-slot="{ currentZone }"
