+++ src/components/Settings/components/Rightside/WalletLimitsCard.vue
@@ -3,35 +3,32 @@
   SPDX-License-Identifier: Apache-2.0
 -->
 <template>
-  <span>
-    <el-card
-      :body-style="{ padding: '0' }"
-      class="settings-card"
-    >
-      <div class="header_btn">
-        <span
-          class="header_btn-title pointed"
-          @click="updateActiveTab(4)"
-        >
-          <span class="header_btn-icon_block">
-            <fa-icon
-              :icon="activeTab === 4 ? 'angle-down' : 'angle-right'"
-              class="header_btn-icon"
-            />
-          </span>
-          Wallet limits
-        </span>
-        <el-button
-          class="action_button"
-          data-cy="addWalletLimit"
-          @click="addWalletLimitDialogVisible = true"
-        >
+  <SettingsCard>
+    <template slot="header">
+      <span
+        class="header_btn-title pointed"
+        @click="updateActiveTab(4)"
+      >
+        <span class="header_btn-icon_block">
           <fa-icon
-            class="action_button-icon"
-            icon="plus"
-          /> Add
-        </el-button>
-      </div>
+            :icon="activeTab === 4 ? 'angle-down' : 'angle-right'"
+            class="header_btn-icon"
+          />
+        </span>
+        Wallet limits
+      </span>
+      <el-button
+        class="action_button"
+        data-cy="addWalletLimit"
+        @click="addWalletLimitDialogVisible = true"
+      >
+        <fa-icon
+          class="action_button-icon"
+          icon="plus"
+        /> Add
+      </el-button>
+    </template>
+    <template slot="content">
       <div v-if="activeTab === 4">
         <div class="settings-item">
           <template v-for="(limit, index) in accountLimits">
@@ -65,200 +62,208 @@
           </div>
         </div>
       </div>
-    </el-card>
-    <el-dialog
-      :visible.sync="addWalletLimitDialogVisible"
-      :show-close="false"
-      data-cy="addWalletLimit"
-      title="Add wallet limit"
-      width="450px"
-      center
-      @closed="onCloseModal"
-    >
-      <div class="approval_form-desc">
-        <el-form
-          ref="limitForm"
-          :model="limitForm"
-          class="dialog-form"
-        >
-          <el-form-item
-            label="Wallets"
-            prop="assetId"
+      <el-dialog
+        :visible.sync="addWalletLimitDialogVisible"
+        :show-close="false"
+        data-cy="addWalletLimit"
+        title="Add wallet limit"
+        width="450px"
+        center
+        @closed="onCloseModal"
+      >
+        <div class="approval_form-desc">
+          <el-form
+            ref="limitForm"
+            :model="limitForm"
+            class="dialog-form"
           >
-            <el-select
-              v-model="limitForm.assetId"
-              class="fullwidth"
+            <el-form-item
+              label="Wallets"
+              prop="assetId"
             >
-              <el-option
-                v-for="wallet in wallets"
-                :key="wallet.name"
-                :label="`${wallet.name} (${wallet.asset.toUpperCase()})`"
-                :value="wallet.assetId"
+              <el-select
+                v-model="limitForm.assetId"
+                class="fullwidth"
+              >
+                <el-option
+                  v-for="wallet in wallets"
+                  :key="wallet.name"
+                  :label="`${wallet.name} (${wallet.asset.toUpperCase()})`"
+                  :value="wallet.assetId"
+                />
+              </el-select>
+            </el-form-item>
+            <el-form-item
+              label="Time"
+              prop="label"
+            >
+              <el-select
+                v-model="limitForm.label"
+                placeholder="Select period"
+                class="fullwidth"
+              >
+                <el-option
+                  v-for="(timelimit, index) in timelimits"
+                  :key="index"
+                  :label="timelimit.label"
+                  :value="timelimit.label"
+                />
+              </el-select>
+            </el-form-item>
+            <el-form-item
+              label="Limit"
+              prop="amount"
+            >
+              <el-input
+                v-model="limitForm.amount"
+                type="number"
+                placeholder="0"
               />
-            </el-select>
-          </el-form-item>
-          <el-form-item
-            label="Time"
-            prop="label"
+            </el-form-item>
+          </el-form>
+        </div>
+        <div
+          slot="footer"
+          class="dialog-form_buttons-block"
+        >
+          <el-button
+            :loading="addingNewLimit"
+            class="dialog-form_buttons action"
+            @click="onAddAssetLimit"
           >
-            <el-select
-              v-model="limitForm.label"
-              placeholder="Select period"
-              class="fullwidth"
-            >
-              <el-option
-                v-for="(timelimit, index) in timelimits"
-                :key="index"
-                :label="timelimit.label"
-                :value="timelimit.label"
-              />
-            </el-select>
-          </el-form-item>
-          <el-form-item
-            label="Limit"
-            prop="amount"
+            Create
+          </el-button>
+          <el-button
+            class="dialog-form_buttons close"
+            @click="addWalletLimitDialogVisible = false"
           >
-            <el-input
-              v-model="limitForm.amount"
-              type="number"
-              placeholder="0"
-            />
-          </el-form-item>
-        </el-form>
-      </div>
-      <div
-        slot="footer"
-        class="dialog-form_buttons-block"
+            Cancel
+          </el-button>
+        </div>
+      </el-dialog>
+      <el-dialog
+        :visible.sync="editWalletLimitDialogVisible"
+        :show-close="false"
+        data-cy="editWalletLimit"
+        title="Edit wallet limit"
+        width="450px"
+        center
       >
-        <el-button
-          :loading="addingNewLimit"
-          class="dialog-form_buttons action"
-          @click="onAddAssetLimit"
-        >
-          Create
-        </el-button>
-        <el-button
-          class="dialog-form_buttons close"
-          @click="addWalletLimitDialogVisible = false"
-        >
-          Cancel
-        </el-button>
-      </div>
-    </el-dialog>
-    <el-dialog
-      :visible.sync="editWalletLimitDialogVisible"
-      :show-close="false"
-      data-cy="editWalletLimit"
-      title="Edit wallet limit"
-      width="450px"
-      center
-    >
-      <div class="approval_form-desc">
-        <el-form
-          :model="editLimitForm"
-          class="dialog-form"
-        >
-          <el-form-item label="Wallets">
-            <el-select
-              v-model="editLimitForm.assetId"
-              class="fullwidth"
-              disabled
-            >
-              <el-option
-                v-for="wallet in wallets"
-                :key="wallet.name"
-                :label="`${wallet.name} (${wallet.asset.toUpperCase()})`"
-                :value="wallet.assetId"
+        <div class="approval_form-desc">
+          <el-form
+            :model="editLimitForm"
+            class="dialog-form"
+          >
+            <el-form-item label="Wallets">
+              <el-select
+                v-model="editLimitForm.assetId"
+                class="fullwidth"
+                disabled
+              >
+                <el-option
+                  v-for="wallet in wallets"
+                  :key="wallet.name"
+                  :label="`${wallet.name} (${wallet.asset.toUpperCase()})`"
+                  :value="wallet.assetId"
+                />
+              </el-select>
+            </el-form-item>
+            <el-form-item label="Time">
+              <el-select
+                v-model="editLimitForm.label"
+                class="fullwidth"
+              >
+                <el-option
+                  v-for="(timelimit, index) in timelimits"
+                  :key="index"
+                  :label="timelimit.label"
+                  :value="timelimit.label"
+                />
+              </el-select>
+            </el-form-item>
+            <el-form-item label="Limit">
+              <el-input
+                v-model="editLimitForm.amount"
+                type="number"
+                placeholder="0"
               />
-            </el-select>
-          </el-form-item>
-          <el-form-item label="Time">
-            <el-select
-              v-model="editLimitForm.label"
-              class="fullwidth"
-            >
-              <el-option
-                v-for="(timelimit, index) in timelimits"
-                :key="index"
-                :label="timelimit.label"
-                :value="timelimit.label"
-              />
-            </el-select>
-          </el-form-item>
-          <el-form-item label="Limit">
-            <el-input
-              v-model="editLimitForm.amount"
-              type="number"
-              placeholder="0"
-            />
-          </el-form-item>
-        </el-form>
-      </div>
-      <div
-        slot="footer"
-        class="dialog-form_buttons-block"
-      >
-        <el-button
-          :loading="editingLimit"
-          class="dialog-form_buttons action"
-          @click="onEditAssetLimit"
+            </el-form-item>
+          </el-form>
+        </div>
+        <div
+          slot="footer"
+          class="dialog-form_buttons-block"
         >
-          Update
-        </el-button>
-        <el-button
-          class="dialog-form_buttons close"
-          @click="editWalletLimitDialogVisible = false"
-        >
-          Cancel
-        </el-button>
-      </div>
-    </el-dialog>
-    <el-dialog
-      :visible.sync="removeAssetLimitVisible"
-      data-cy="removeLimitDialog"
-      title="Asset limit"
-      width="450px"
-      center
-    >
-      <div
-        v-if="limitToRemove.wallet"
-        class="approval_form-desc"
+          <el-button
+            :loading="editingLimit"
+            class="dialog-form_buttons action"
+            @click="onEditAssetLimit"
+          >
+            Update
+          </el-button>
+          <el-button
+            class="dialog-form_buttons close"
+            @click="editWalletLimitDialogVisible = false"
+          >
+            Cancel
+          </el-button>
+        </div>
+      </el-dialog>
+      <el-dialog
+        :visible.sync="removeAssetLimitVisible"
+        data-cy="removeLimitDialog"
+        title="Asset limit"
+        width="450px"
+        center
       >
-        Are you sure want to remove limit for <b>{{ limitToRemove.wallet.name }} ({{ limitToRemove.wallet.asset }})</b> ?
-      </div>
-      <div
-        slot="footer"
-        class="dialog-form_buttons-block"
-      >
-        <el-button
-          :loading="removingLimit"
-          class="dialog-form_buttons action"
-          type="danger"
-          @click="onRemoveAssetLimit"
-        >Remove
-        </el-button>
-        <el-button
-          class="dialog-form_buttons close"
-          @click="removeAssetLimitVisible = false"
+        <div
+          v-if="limitToRemove.wallet"
+          class="approval_form-desc"
         >
-          Cancel
-        </el-button>
-      </div>
-    </el-dialog>
-  </span>
+          Are you sure want to remove limit for <b>{{ limitToRemove.wallet.name }} ({{ limitToRemove.wallet.asset }})</b> ?
+        </div>
+        <div
+          slot="footer"
+          class="dialog-form_buttons-block"
+        >
+          <el-button
+            :loading="removingLimit"
+            class="dialog-form_buttons action"
+            type="danger"
+            @click="onRemoveAssetLimit"
+          >Remove
+          </el-button>
+          <el-button
+            class="dialog-form_buttons close"
+            @click="removeAssetLimitVisible = false"
+          >
+            Cancel
+          </el-button>
+        </div>
+      </el-dialog>
+    </template>
+  </SettingsCard>
 </template>
 
 <script>
 import { mapGetters, mapActions } from 'vuex'
+import { lazyComponent } from '@router'
 
 export default {
   name: 'WalletLimitsCard',
+  components: {
+    SettingsCard: lazyComponent('Settings/components/Rightside/SettingsCard')
+  },
   props: {
     activeTab: {
       type: Number,
       required: true
     },
     updateActiveTab: {
+      type: Function,
+      required: true
+    },
+    openApprovalDialog: {
       type: Function,
       required: true
     }
@@ -307,9 +312,11 @@ export default {
       'wallets'
     ])
   },
+  created () {
+    this.getAccountLimits()
+  },
   methods: {
     ...mapActions([
-      'openApprovalDialog',
       'updateAccount',
       'addAssetLimit',
       'removeAssetLimit',
@@ -442,13 +449,13 @@ export default {
   font-size: 1rem;
   border-radius: 0.1rem;
 }
- .dialog-form >>> .el-input__inner:focus {
+.dialog-form >>> .el-input__inner:focus {
   opacity: 1;
 }
- .dialog-form >>> .el-form-item {
+.dialog-form >>> .el-form-item {
   height: 4.4rem;
 }
- .dialog-form >>> .el-form-item__label {
+.dialog-form >>> .el-form-item__label {
   line-height: 1;
   position: relative;
   top: 1.5rem;
